+++ src/application/containers/views/sources/graphql/queries/allDatabases.graphql
