+++ src/application/containers/views/sources/graphql/queries/computeDatabaseMappingProgress.graphql
@@ -1,0 +1,3 @@
+query computeDatabaseMappingProgress($databaseId: ID!) {
+    computeDatabaseMappingProgress(databaseId: $databaseId)
+}
