+++ src/application/containers/views/sources/style.less
@@ -7,6 +7,10 @@ h2 {
     margin-top: 0;
 }
 
+#main-container-softwares {
+    padding: @window-padding;
+}
+
 #software-cards {
     display: flex;
     flex-direction: row;
